requirements.txt
tensorflow
matplotlib
fastapi
uvicorn
pillow
